Release 1.3.11.beta.1

+++ Entity/StoreStack/MtdStoreStackFile.cs
@@ -18,6 +18,8 @@
 */
 
 
+using System.Configuration;
+
 namespace Mtd.OrderMaker.Server.Entity
 {
     public partial class MtdStoreStackFile
@@ -25,7 +27,7 @@ namespace Mtd.OrderMaker.Server.Entity
         public long Id { get; set; }
         public byte[] Register { get; set; }
         public string FileName { get; set; }
-        public int FileSize { get; set; }
+        public long FileSize { get; set; }
         public string FileType { get; set; }
 
         public virtual MtdStoreStack IdNavigation { get; set; }
